+++ src/lib.rs
@@ -56,12 +56,13 @@ use bevy_ecs::prelude::*;
 use bevy_render::{RenderApp, RenderStage};
 pub use bevy_sprite::ColorMaterial;
 use bevy_sprite::SpriteSystem::ExtractSprites;
-pub use bundle::SpriteBundle;
-pub use sprite::Sprite;
+
+pub use {bundle::SpriteBundle, sprite::Sprite, texture_atlas::TextureAtlasSprite};
 
 mod bundle;
 mod extract;
 mod sprite;
+mod texture_atlas;
 
 /// Plugin to use sprites with materials.
 ///