+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Bevy `0.7`
 * `Sprite` supports `bevy_sprite::Anchor`
+* `TextureAtlasSprite` support
+* using `copyless` crate
 
 ## 0.1.0